ajustes de dependencias quebrando, e mudança de tipo de atributo de Integer para BigDecimal

+++ src/main/java/br/com/ada/reservala/controller/dto/RoomDTOResponse.java
@@ -4,14 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @Builder
 public class RoomDTOResponse {
-
     private Integer roomNumber;
     private String type;
-    private Integer price;
+    private BigDecimal price;
     private Boolean avalaible;
 
 }
